fix segfault with insc after lenc due to use of int instead of long; add INCLUDES to mkdis so packages outside the distribution can use their own module folders

+++ libinterp/string.c
@@ -301,7 +301,7 @@ OP(cvtac)
 
 OP(lenc)
 {
-	int l;
+	ulong l;
 	String *ss;
 
 	l = 0;
